# Attempt #2: always allocate for the maximum possible digest output