--- a/blake2s/blake2s.go
+++ b/blake2s/blake2s.go
@@ -313,14 +313,14 @@ func (d *Digest) finalize(out []byte) error {
 
 	dCopy.compress()
 
-	var shift uint
-	var mask uint32
-
-	for offset := 0; offset < len(out); offset++ {
-		shift = 8 * (uint(offset) % 4)
-		mask = uint32(0xFF << shift)
-		out[offset] = byte((dCopy.h[offset/4] & mask) >> shift)
-	}
+	putU32LE(out[0*4:], dCopy.h[0])
+	putU32LE(out[1*4:], dCopy.h[1])
+	putU32LE(out[2*4:], dCopy.h[2])
+	putU32LE(out[3*4:], dCopy.h[3])
+	putU32LE(out[4*4:], dCopy.h[4])
+	putU32LE(out[5*4:], dCopy.h[5])
+	putU32LE(out[6*4:], dCopy.h[6])
+	putU32LE(out[7*4:], dCopy.h[7])
 
 	return nil
 }
@@ -423,7 +423,7 @@ func (d *Digest) Write(input []byte) (n int, err error) {
 // It does not change the underlying hash state.
 func (d *Digest) Sum(b []byte) (out []byte) {
 	// if there's space, reuse the b slice
-	if n := len(b) + d.size; cap(b) >= n {
+	if n := len(b) + 32; cap(b) >= n {
 		out = b[:n]
 	} else {
 		out = make([]byte, n)
@@ -436,7 +436,7 @@ func (d *Digest) Sum(b []byte) (out []byte) {
 		return out[:len(b)]
 	}
 
-	return out
+	return out[:len(b)+d.size]
 }
 
 // Reset resets the Hash to its initial state.