This alternative approach notably improves the performance on small input tasks (like hashing 32-byte keys or group elements), mitigating most of the penalty of the earlier fix at the cost of unintuitive allocation behavior.

cpu: Intel(R) Core(TM) i7-7560U CPU @ 2.40GHz
go version devel +b0f01e17f8 Wed Dec 16 22:45:19 2020 +0000 linux/amd64

blake2s:
```
name          old time/op    new time/op    delta
Hash8Bytes-4     345ns ± 2%     293ns ± 1%  -15.07%  (p=0.000 n=10+9)
Hash1K-4        2.81µs ± 0%    2.76µs ± 1%   -1.86%  (p=0.000 n=8+9)
Hash8K-4        20.7µs ± 1%    20.5µs ± 1%   -1.02%  (p=0.000 n=9+9)

name          old speed      new speed      delta
Hash8Bytes-4  23.2MB/s ± 2%  27.3MB/s ± 1%  +17.74%  (p=0.000 n=10+9)
Hash1K-4       364MB/s ± 0%   371MB/s ± 1%   +1.90%  (p=0.000 n=8+9)
Hash8K-4       396MB/s ± 1%   400MB/s ± 1%   +1.03%  (p=0.000 n=9+9)
```
blake2b:
```
name          old time/op    new time/op    delta
Hash8Bytes-4     459ns ± 1%     366ns ± 1%  -20.40%  (p=0.000 n=10+10)
Hash1K-4        1.87µs ± 0%    1.79µs ± 0%   -4.36%  (p=0.000 n=10+10)
Hash8K-4        13.0µs ± 1%    13.0µs ± 0%     ~     (p=0.113 n=10+9)

name          old speed      new speed      delta
Hash8Bytes-4  17.4MB/s ± 1%  21.9MB/s ± 1%  +25.62%  (p=0.000 n=10+10)
Hash1K-4       547MB/s ± 0%   572MB/s ± 0%   +4.55%  (p=0.000 n=10+10)
Hash8K-4       629MB/s ± 1%   628MB/s ± 0%     ~     (p=0.113 n=10+9)
```

cc @elichai @filosottile for thoughts on if this is "better" that the prior approach